+++ R/shapviz.R
@@ -5,7 +5,7 @@
 #' from a fitted model of type
 #' - XGBoost,
 #' - LightGBM, or
-#' - H2O (tree-based regression or binary classification model).
+#' - H2O.
 #'
 #' Furthermore, [shapviz()] can digest the results of
 #' - `fastshap::explain()`,
@@ -116,7 +116,7 @@ shapviz.matrix = function(
 #' # XGBoost models
 #' X_pred <- data.matrix(iris[, -1])
 #' dtrain <- xgboost::xgb.DMatrix(X_pred, label = iris[, 1], nthread = 1)
-#' fit <- xgboost::xgb.train(data = dtrain, nrounds = 10, nthread = 1)
+#' fit <- xgboost::xgb.train(list(nthread = 1), data = dtrain, nrounds = 10)
 #'
 #' # Will use numeric matrix "X_pred" as feature matrix
 #' x <- shapviz(fit, X_pred = X_pred)
@@ -131,12 +131,12 @@ shapviz.matrix = function(
 #' x <- shapviz(fit, X_pred = dtrain, X = iris)
 #'
 #' # Multiclass setting
-#' params <- list(objective = "multi:softprob", num_class = 3)
+#' params <- list(objective = "multi:softprob", num_class = 3, nthread = 1)
 #' X_pred <- data.matrix(iris[, -5])
 #' dtrain <- xgboost::xgb.DMatrix(
 #'   X_pred, label = as.integer(iris[, 5]) - 1, nthread = 1
 #' )
-#' fit <- xgboost::xgb.train(params = params, data = dtrain, nrounds = 10, nthread = 1)
+#' fit <- xgboost::xgb.train(params = params, data = dtrain, nrounds = 10)
 #'
 #' # Select specific class
 #' x <- shapviz(fit, X_pred = X_pred, which_class = 3)
@@ -149,7 +149,7 @@ shapviz.matrix = function(
 #' # What if we would have one-hot-encoded values and want to explain the original column?
 #' X_pred <- stats::model.matrix(~ . -1, iris[, -1])
 #' dtrain <- xgboost::xgb.DMatrix(X_pred, label = as.integer(iris[, 1]), nthread = 1)
-#' fit <- xgboost::xgb.train(data = dtrain, nrounds = 10, nthread = 1)
+#' fit <- xgboost::xgb.train(list(nthread = 1), data = dtrain, nrounds = 10)
 #' x <- shapviz(
 #'   fit,
 #'   X_pred = X_pred,
@@ -410,10 +410,10 @@ shapviz.predict_parts <- function(object, ...) {
 #' @describeIn shapviz
 #'   Creates a "shapviz" object from `shapr::explain()`.
 #' @export
-shapviz.shapr <- function(object, X = object[["x_test"]], collapse = NULL, ...) {
-  dt <- as.matrix(object[["dt"]])
+shapviz.shapr <- function(object, X = as.data.frame(object$internal$data$x_explain), collapse = NULL, ...) {
+  dt <- as.matrix(object[["shapley_values_est"]])
   shapviz.matrix(
-    object = dt[, setdiff(colnames(dt), "none"), drop = FALSE],
+    object = dt[, setdiff(colnames(dt), c("none","explain_id")), drop = FALSE],
     X = X,
     baseline = dt[1L, "none"],
     collapse = collapse
@@ -454,28 +454,26 @@ shapviz.kernelshap <- function(
 }
 
 #' @describeIn shapviz
-#'   Creates a "shapviz" object from a (tree-based) H2O regression model.
-#' @export
-shapviz.H2ORegressionModel = function(
-    object, X_pred, X = as.data.frame(X_pred), collapse = NULL, ...
-  ) {
-  shapviz.H2OModel(object = object, X_pred = X_pred, X = X, collapse = collapse, ...)
-}
-
-#' @describeIn shapviz
-#'   Creates a "shapviz" object from a (tree-based) H2O binary classification model.
-#' @export
-shapviz.H2OBinomialModel = function(
-    object, X_pred, X = as.data.frame(X_pred), collapse = NULL, ...
-  ) {
-  shapviz.H2OModel(object = object, X_pred = X_pred, X = X, collapse = collapse, ...)
-}
-
-#' @describeIn shapviz
-#'   Creates a "shapviz" object from a (tree-based) H2O model (base class).
+#'   Creates a "shapviz" object from an H2O model.
+#' @param background_frame Background dataset for baseline SHAP or marginal SHAP.
+#'    Only for H2O models.
+#' @param output_space If model has link function, this argument controls whether the
+#'   SHAP values should be linearly (= approximately) transformed to the original scale
+#'   (if `TRUE`). The default is to return the values on link scale.
+#'   Only for H2O models.
+#' @param output_per_reference Switches between different algorithms, see
+#'   `?h2o::h2o.predict_contributions` for details.
+#'   Only for H2O models.
 #' @export
 shapviz.H2OModel = function(
-    object, X_pred, X = as.data.frame(X_pred), collapse = NULL, ...
+    object,
+    X_pred,
+    X = as.data.frame(X_pred),
+    collapse = NULL,
+    background_frame = NULL,
+    output_space = FALSE,
+    output_per_reference = FALSE,
+    ...
   ) {
   if (!requireNamespace("h2o", quietly = TRUE)) {
     stop("Package 'h2o' not installed")
@@ -488,7 +486,16 @@ shapviz.H2OModel = function(
   if (!inherits(X_pred, "H2OFrame")) {
     X_pred <- h2o::as.h2o(X_pred)
   }
-  S <- as.matrix(h2o::h2o.predict_contributions(object, newdata = X_pred, ...))
+  S <- as.matrix(
+    h2o::h2o.predict_contributions(
+      object,
+      newdata = X_pred,
+      background_frame = background_frame,
+      output_space = output_space,
+      output_per_reference = output_per_reference,
+      ...
+    )
+  )
   shapviz.matrix(
     object = S[, setdiff(colnames(S), "BiasTerm"), drop = FALSE],
     X = X,